Base Functionalities of Mechanics 1 - 5
According to GDD.

There are 3 types of objects currently - Character, Projectile, Item. Of which extends themselves from a base template of

CharacterTemplate, ProjectileTemplate, and ItemTemplate respectively.

Most of the functions_ are executed from the templates, while more specific things (like a special power, etc) can be executed from the specific object.tscn

Current implementations are a basic WASD-movement, LMB and RMB mouse-button projectile shooting, and basic item-drop function (currently, the CharacterHuman triggers this).

===== RESOURCE LOADING LOGIC ======

There are currently two Autoload scripts called ProjectilesPreloader and ItemsPreloader. Each of them are from a .tscn with a single ResourcePreloader.node . Note that the single nodes being a ResourcePreloader currently makes no difference whatsoever. It is the script that does the work.

An object, per example > a Human who dies ( _die() ), makes a call to the relevant preloader by calling for a function inside of their script, and gets returned a preload() resource. From there, it's a standard Godot load/add_child code.